adapters/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zfw-heimdall-python"
version = "0.1.0"
description = "Python adapter for Heimdall Error Orchestration"
dependencies = [
    "fastapi",
    "starlette",
    "ulid-py",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["heimdall*"]
